Rename to @tobius/istruthy
There is a competing project published to npmjs.com/is-truthy (which is
too similar to this project name to allow this to be published). That
project is 3 years old and does not actually compute anything, it simply
mirrors true/false results for native types. This project actually
accounts for values that a human may "think" is "truthy" even when a
native type test says otherwise.